Move files around; set default profile image to canvas logo
